Creación Informe Utilización de Cargadores en todas las capas

+++ ProgramacionTP_CS_API_PostgreSQL_Dapper/ProgramacionTP_CS_API_PostgreSQL_Dapper/Models/InformeOperacionAutobus.cs
@@ -3,6 +3,6 @@
     public class InformeOperacionAutobus
     {
         public int Hora { get; set; } = 0;
-        public float Total_operacion_autobuses { get; set; } = 0;
+        public int Total_operacion_autobuses { get; set; } = 0;
     }
 }
